Set version automatically
Convert the check script to a script that sets the version to match
whatever Riot is being packaged.

Add a README because the two ways of interacting with versions are
starting to get complex to reason about.

Also add a docker build script.
