+++ package.json
@@ -14,10 +14,10 @@
   "scripts": {
     "mkdirs": "mkdirp packages deploys",
     "fetch": "yarn run mkdirs && node scripts/fetch-package.js",
-    "check": "node scripts/check-webapp.js",
+    "setversion": "node scripts/set-version.js",
     "start": "electron .",
     "lint": "eslint src/",
-    "build": "yarn run check && electron-builder",
+    "build": "yarn run setversion && electron-builder",
     "clean": "rimraf webapp.asar dist packages deploys"
   },
   "dependencies": {
